Using another version of ym2203 engine.
git-svn-id: https://zxtune.googlecode.com/svn/trunk@1779 e19a5227-2556-0410-816f-a35683127991

+++ src/devices/fm/fm.cpp
@@ -23,7 +23,7 @@ namespace
 {
   using namespace Devices::FM;
 
-  typedef int YM2203SampleType;
+  typedef short YM2203SampleType;
 
   uint_t GetBandByFreq(uint_t freq)
   {
